requirements.txt
flask
flask-login
flask-sqlalchemy
flask-bcrypt
flask-wtf
flask-cors
cozepy
python-dotenv 